+++ Entity.h
@@ -5,31 +5,31 @@ public:
 	Entity();
 	~Entity();
 
-	void Init(int posx, int posy, int w, int h, int s);
-	void GetRect(int *posx, int *posy, int *w, int *h);
-	void SetX(int posx);
-	void SetY(int posy);
-	void SetDX(int posx);
-	void SetDY(int posy);
-	void SetSpeed(int speed);
-	int  GetX();
-	int  GetY();
-	int  GetDX();
-	int  GetDY();
-	int  GetWidth();
-	int  GetHeight();
-	int  GetSpeed();
+	void Init(float posx, float posy, float w, float h, float s);
+	void GetRect(int *posx, int*posy, int*w, int*h);
+	void SetX(float posx);
+	void SetY(float posy);
+	void SetDX(float posx);
+	void SetDY(float posy);
+	void SetSpeed(float speed);
+	float  GetX();
+	float  GetY();
+	float  GetDX();
+	float  GetDY();
+	float  GetWidth();
+	float  GetHeight();
+	float  GetSpeed();
 	void ShutDown(); 
 	bool IsAlive();
 	void Move();
-	void SetDir(int dx, int dy);
+	void SetDir(float dx, float dy);
 	bool Collide(Entity other);
 
 
 private:
-	int x, y, dx, dy;
-	int width, height;
-	int speed;
+	float x, y, dx, dy;
+	float width, height;
+	float speed;
 	bool is_alive;
 };
